make_tarball: workaround for broken tar on Darwin
On OS X /usr/bin/tar bundles extended attributes that trigger warnings
when the tarball is unpacked with GNU tar.
See http://cpants.cpanauthors.org/kwalitee/no_pax_headers

So when the tarball is made using a tar executable on Darwin we take
care of setting the appropriate environment variables to avoid the broken
behaviour. The fix is the same as in ExtUtils::MM_Darwin.